Fix Rubocop by adding in shared style guide. Regen todo - This commit adds in the TED shared rubcop style guide. Details on this   guide can be found [here](https://github.com/tedconf/code-style-guides/tree/master/linters/rubocop)
In the process of adding in this style guide, I then regenerated the
.rubocop_todo.yml file to ignore the current violation that are in this
gem.

Details on rubocop, its configuration, and the .rubocop_todo.yml can be
found in the [Rubocop
Documentaion](https://rubocop.readthedocs.io/en/latest/configuration/)

At this time, I am currently ignoring all rubocop violations. The reason
behind ignoring all violations is to make this a valid jenkins project,
so that we start to receive the benefits of continual testing. The
reasoning is not malice, incompetence, and/or laziness, as the plan is
to come back and fix the violations in smaller, more controlled batches.

I feel that rubocop is indeed a very valued addition to our workflow,
however at this time, the benefits of rubocop are only really truly felt
when the project is integrated into jenkins, and that is why that is a
higher priority at this time.

tl;dr FR